+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 	Placeholder for the next version (at the beginning of the line):
 	## __WORK IN PROGRESS__
 -->
+
+## **WORK IN PROGRESS**
+
+-   [#165](https://github.com/MiSchroe/klf-200-api/issues/165) Fix logging error when the login throws an error in the TLS stack.
+-   Fix lint findings.
+
 ## 5.0.1 (2024-07-17)
 
 -   [#147](https://github.com/MiSchroe/klf-200-api/issues/147) refreshLimitationStatusAsync should wait for the session finished notification.